Move TransactionNotificationsService to service package
